Extract complex method in /glyph
